fix(relativeTime): prevent relativeTime to add resources to soon
This may prevent translation to be loaded from other sources. For
instance, the XHR backend uses hasResourceBundle to decide whether to
load resources or not. But the call to i18next.addResources in
relativeTime makes this function always return true. Thus, the
translations of the user are never loaded.

This commit delays the execution of addResources in relativeTime in
order for both translations to be loaded correctly.