import was using incorrect syntax for typescript
Typescript doesn't allow this "default" import syntax for CommonJs modules, so you have to use the old CommonJs syntax to get it to import. The following will also work:

import * as i18next from 'i18next';

But I prefer the older syntax so that it is super clear that this is a compatibility issue.